service,db: use new common types
